requirements.txt
google-auth~=2.30.0
google-auth-oauthlib~=1.2.0
google-auth-httplib2~=0.2.0
google-api-python-client~=2.132.0
ics~=0.7.2